requirements.txt
PyYAML==3.12
psutil==5.6.3
pytest==3.1.3
pyreadline==2.1
requests==2.18.1
pyftpdlib==1.5.3
pexpect==4.4.0
matplotlib==2.2.4
ipaddress==1.0.22
enum34==1.1.6
crcmod==1.7
terminaltables==3.1.0
pyserial==3.0
scapy==2.4.3
sounddevice==0.3.14

pcapy==0.11.1; sys_platform != 'win32'

brainstem==2.7.8; sys_platform == 'linux2'

WMI==1.4.9; sys_platform == 'win32'
pywinauto==0.6.1; sys_platform == 'win32'
